pyproject.toml
[tool.poetry]
name = "weiroll-py"
version = "0.2.0"
description = "Build weiroll transactions with brownie"
authors = ["FP <FP@no-email.com>"]
license = "MIT"
packages = [
    { include = "./weiroll.py" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
eth-brownie = "^v1.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
